+++ rig-core/src/agent/mod.rs
@@ -112,10 +112,10 @@ pub(crate) mod prompt_request;
 mod tool;
 
 pub use crate::message::Text;
-pub use builder::AgentBuilder;
+pub use builder::{AgentBuilder, AgentBuilderSimple};
 pub use completion::Agent;
 pub use prompt_request::PromptHook;
 pub use prompt_request::streaming::{
     FinalResponse, MultiTurnStreamItem, StreamingPromptRequest, stream_to_stdout,
 };
-pub use prompt_request::{PromptRequest, PromptResponse};
+pub use prompt_request::{CancelSignal, PromptRequest, PromptResponse};