+++ rig-core/src/completion/request.rs
@@ -64,9 +64,11 @@
 //! the individual traits, structs, and enums defined in this module.
 
 use super::message::{AssistantContent, DocumentMediaType};
+use crate::client::builder::FinalCompletionResponse;
 use crate::client::completion::CompletionModelHandle;
 use crate::message::ToolChoice;
 use crate::streaming::StreamingCompletionResponse;
+use crate::tool::server::ToolServerError;
 use crate::wasm_compat::{WasmBoxedFuture, WasmCompatSend, WasmCompatSync};
 use crate::{OneOrMany, http_client, streaming};
 use crate::{
@@ -126,6 +128,10 @@ pub enum PromptError {
     #[error("ToolCallError: {0}")]
     ToolError(#[from] ToolSetError),
 
+    /// There was an issue while executing a tool on a tool server
+    #[error("ToolServerError: {0}")]
+    ToolServerError(#[from] ToolServerError),
+
     /// The LLM tried to call too many tools during a multi-turn conversation.
     /// To fix this, you may either need to lower the amount of tools your model has access to (and then create other agents to share the tool load)
     /// or increase the amount of turns given in `.multi_turn()`.
@@ -135,6 +141,18 @@ pub enum PromptError {
         chat_history: Box<Vec<Message>>,
         prompt: Message,
     },
+
+    /// A prompting loop was cancelled.
+    #[error("PromptCancelled")]
+    PromptCancelled { chat_history: Box<Vec<Message>> },
+}
+
+impl PromptError {
+    pub(crate) fn prompt_cancelled(chat_history: Vec<Message>) -> Self {
+        Self::PromptCancelled {
+            chat_history: Box::new(chat_history),
+        }
+    }
 }
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
@@ -167,7 +185,7 @@ impl std::fmt::Display for Document {
     }
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
 pub struct ToolDefinition {
     pub name: String,
     pub description: String,
@@ -363,7 +381,10 @@ pub trait CompletionModelDyn: WasmCompatSend + WasmCompatSync {
     fn stream(
         &self,
         request: CompletionRequest,
-    ) -> WasmBoxedFuture<'_, Result<StreamingCompletionResponse<()>, CompletionError>>;
+    ) -> WasmBoxedFuture<
+        '_,
+        Result<StreamingCompletionResponse<FinalCompletionResponse>, CompletionError>,
+    >;
 
     fn completion_request(
         &self,
@@ -394,7 +415,10 @@ where
     fn stream(
         &self,
         request: CompletionRequest,
-    ) -> WasmBoxedFuture<'_, Result<StreamingCompletionResponse<()>, CompletionError>> {
+    ) -> WasmBoxedFuture<
+        '_,
+        Result<StreamingCompletionResponse<FinalCompletionResponse>, CompletionError>,
+    > {
         Box::pin(async move {
             let resp = self.stream(request).await?;
             let inner = resp.inner;