+++ rig-core/src/lib.rs
@@ -110,15 +110,19 @@ pub mod agent;
 #[cfg(feature = "audio")]
 #[cfg_attr(docsrs, doc(cfg(feature = "audio")))]
 pub mod audio_generation;
-pub mod cli_chatbot;
 pub mod client;
 pub mod completion;
 pub mod embeddings;
+
+#[cfg(feature = "experimental")]
+#[cfg_attr(docsrs, doc(cfg(feature = "experimental")))]
+pub mod evals;
 pub mod extractor;
 pub mod http_client;
 #[cfg(feature = "image")]
 #[cfg_attr(docsrs, doc(cfg(feature = "image")))]
 pub mod image_generation;
+pub mod integrations;
 pub(crate) mod json_utils;
 pub mod loaders;
 pub mod one_or_many;