+++ rig-sqlite/tests/integration_test.rs
@@ -1,4 +1,4 @@
-use rig::vector_store::request::VectorSearchRequest;
+use rig::vector_store::request::{SearchFilter, VectorSearchRequest};
 use serde_json::json;
 
 use rig::client::EmbeddingsClient;
@@ -8,7 +8,9 @@ use rig::{
     embeddings::{Embedding, EmbeddingsBuilder},
     providers::openai,
 };
-use rig_sqlite::{Column, ColumnValue, SqliteVectorStore, SqliteVectorStoreTable};
+use rig_sqlite::{
+    Column, ColumnValue, SqliteSearchFilter, SqliteVectorStore, SqliteVectorStoreTable,
+};
 use rusqlite::ffi::{sqlite3, sqlite3_api_routines, sqlite3_auto_extension};
 use sqlite_vec::sqlite3_vec_init;
 use tokio_rusqlite::Connection;
@@ -49,6 +51,143 @@ type SqliteExtensionFn =
 
 #[tokio::test]
 async fn vector_search_test() {
+    // Initialize the `sqlite-vec`extension
+    // See: https://alexgarcia.xyz/sqlite-vec/rust.html
+    unsafe {
+        sqlite3_auto_extension(Some(std::mem::transmute::<*const (), SqliteExtensionFn>(
+            sqlite3_vec_init as *const (),
+        )));
+    }
+
+    // Initialize SQLite connection
+    let conn = Connection::open("vector_store.db")
+        .await
+        .expect("Could not initialize SQLite connection");
+
+    // Setup mock openai API
+    let server = httpmock::MockServer::start();
+
+    server.mock(|when, then| {
+        when.method(httpmock::Method::POST)
+            .path("/embeddings")
+            .header("Authorization", "Bearer TEST")
+            .json_body(json!({
+                "input": [
+                    "Definition of a *flurbo*: A flurbo is a green alien that lives on cold planets",
+                    "Definition of a *glarb-glarb*: A glarb-glarb is a ancient tool used by the ancestors of the inhabitants of planet Jiro to farm the land.",
+                    "Definition of a *linglingdong*: A term used by inhabitants of the far side of the moon to describe humans."
+                ],
+                "model": "text-embedding-ada-002",
+                "dimensions": 1536,
+            }));
+        then.status(200)
+            .header("content-type", "application/json")
+            .json_body(json!({
+                "object": "list",
+                "data": [
+                  {
+                    "object": "embedding",
+                    "embedding": vec![-0.001; 1536],
+                    "index": 0
+                  },
+                  {
+                    "object": "embedding",
+                    "embedding": vec![0.0023064255; 1536],
+                    "index": 1
+                  },
+                  {
+                    "object": "embedding",
+                    "embedding": vec![-0.001; 1536],
+                    "index": 2
+                  },
+                ],
+                "model": "text-embedding-ada-002",
+                "usage": {
+                  "prompt_tokens": 8,
+                  "total_tokens": 8
+                }
+            }
+        ));
+    });
+    server.mock(|when, then| {
+        when.method(httpmock::Method::POST)
+            .path("/embeddings")
+            .header("Authorization", "Bearer TEST")
+            .json_body(json!({
+                "input": [
+                    "What is a glarb?",
+                ],
+                "model": "text-embedding-ada-002",
+                "dimensions": 1536,
+            }));
+        then.status(200)
+            .header("content-type", "application/json")
+            .json_body(json!({
+                    "object": "list",
+                    "data": [
+                      {
+                        "object": "embedding",
+                        "embedding": vec![0.0024064254; 1536],
+                        "index": 0
+                      }
+                    ],
+                    "model": "text-embedding-ada-002",
+                    "usage": {
+                      "prompt_tokens": 8,
+                      "total_tokens": 8
+                    }
+                }
+            ));
+    });
+
+    // Initialize OpenAI client
+    let openai_client = openai::Client::builder("TEST")
+        .base_url(&server.base_url())
+        .build();
+
+    // Select the embedding model and generate our embeddings
+    let model = openai_client.embedding_model(openai::TEXT_EMBEDDING_ADA_002);
+
+    let embeddings = create_embeddings(model.clone()).await;
+
+    // Initialize SQLite vector store
+    let vector_store = SqliteVectorStore::new(conn, &model)
+        .await
+        .expect("Could not initialize SQLite vector store");
+
+    // Add embeddings to vector store
+    vector_store
+        .add_rows(embeddings)
+        .await
+        .expect("Could not add embeddings to vector store");
+
+    // Create a vector index on our vector store
+    let index = vector_store.index(model);
+    let query = "What is a glarb?";
+    let samples = 1;
+    let req = VectorSearchRequest::builder()
+        .samples(samples)
+        .query(query)
+        .build()
+        .expect("VectorSearchRequest should not fail to build here");
+
+    // Query the index
+    let results = index.top_n::<serde_json::Value>(req).await.expect("");
+
+    let (_, _, value) = &results.first().expect("");
+
+    assert_eq!(
+        value,
+        &serde_json::json!({
+            "id": "doc1",
+            "definition": "Definition of a *glarb-glarb*: A glarb-glarb is a ancient tool used by the ancestors of the inhabitants of planet Jiro to farm the land.",
+        })
+    )
+}
+
+// TODO: (@FayCarsons) This doesn't really test much
+#[tokio::test]
+async fn vector_search_filter_test() {
     // Initialize the `sqlite-vec`extension
     // See: https://alexgarcia.xyz/sqlite-vec/rust.html
     unsafe {
@@ -164,21 +303,13 @@ async fn vector_search_test() {
     let req = VectorSearchRequest::builder()
         .samples(samples)
         .query(query)
+        .filter(SqliteSearchFilter::eq("id".into(), &"doc1").not())
         .build()
         .expect("VectorSearchRequest should not fail to build here");
 
     // Query the index
     let results = index.top_n::<serde_json::Value>(req).await.expect("");
-
-    let (_, _, value) = &results.first().expect("");
-
-    assert_eq!(
-        value,
-        &serde_json::json!({
-            "id": "doc1",
-            "definition": "Definition of a *glarb-glarb*: A glarb-glarb is a ancient tool used by the ancestors of the inhabitants of planet Jiro to farm the land.",
-        })
-    )
+    assert!(results.is_empty());
 }
 
 async fn create_embeddings(model: openai::EmbeddingModel) -> Vec<(Word, OneOrMany<Embedding>)> {